Unignored (for git) some parts of the project and renamed the external
Files concerned (added to the rep):
- source code of the min-api and of the external.
- externals folder containing the released external
- help folder containing the Max help patch (placeholder)
- tests folder containing some build tests (not working)
- package-info.json is a description of the package

Renamed the external to gexture_pkg.gexture_tilde as input will be signal.
Build (for Win10-VS) folder stays ignored as it's system specific and contains
a lot of files.
Project build can be done with 'cmake ..'.
Release build can be done with 'cmake --build . --config Release'.